Fix styling Settings Dialog
